requirements.txt
Flask
Flask-Cors
Flask-Login
Flask-PyMongo
pymongo
requests
python-dotenv
